Nav bar in satisfactory state
Nav bar is in a satisfactory form for now. Will move on to other
elements. Plan is to add greater sophistication later.

Some annoyances with plain CSS. It fails silently; even on a syntax
error like a missing semicolon won't come up in either Fresh's console
or the browser inspector. Even if there was a build step to catch this,
a naming error (PageNave instead of PageNav) might still be missed. I'm
sure there's a linter which could pick up on this.

Will take quite a while before this app is impressive. Over time this
will become a great asset, but real webdevs would use more tooling to
increase productivity. Might switch over to an even simpler, more
cookie-cutter app to get something on CV.