Video Page starting mock
Begun mocking up the video page. Did a little refactoring around the
edges as well.

It occurred to me I should be taking greater advantage of native HTML
elements, to avoid just creating a div-css soup. Of course I should
remember not to fall into the trap of premature abstraction, and to
leave things be until changing them solves a problem.

Tweaked _app.tsx slightly. It now expects all components to come wrapped
in a <main> element. I figure since there needs to be a single top-level
element anyway, this is good practice.