Continued work on Video Page
Continued work on the video page, was a bit distracted today so I didn't
get much done. Just positioned the mock video and the title.

One thing I am keeping in mind moving forward is semantic HTML. The
mock video is in an `article`, the title in a `footer` of that
`article`. Important for SEO and is an alternative to `div` soup.